changed all internal math to doubles

+++ src/vector.c
@@ -1,7 +1,7 @@
 #include <math.h>
 #include "vector.h"
 
-float vec3dot(const vec3 v1, const vec3 v2)
+double vec3dot(const vec3 v1, const vec3 v2)
 {
     return v1.x * v2.x + v1.y * v2.y + v1.z * v2.z;
 }
@@ -9,7 +9,7 @@ float vec3dot(const vec3 v1, const vec3 v2)
 vec3 vec3normalize(const vec3 v1)
 {
     vec3 out = {0};
-    float len = sqrt(v1.x * v1.x + v1.y * v1.y + v1.z * v1.z);
+    double len = sqrt(v1.x * v1.x + v1.y * v1.y + v1.z * v1.z);
 
     if(len != 0)
     {
@@ -48,7 +48,7 @@ vec3 vec3sub(const vec3 v1, const vec3 v2)
     return out;
 }
 
-vec3 vec3mul(const vec3 v1, const float s)
+vec3 vec3mul(const vec3 v1, const double s)
 {
     vec3 out;
     out.x = v1.x * s;
